requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==3.1.2
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.6.2
Werkzeug==3.1.4
